Some changes to the 2D boundary conditions for pure CG pairs.
Make sure the correct boundary terms are applied for the case where
the pressure gradient and the divergence term are *not* integrated by
parts. This a.o. enables Taylor Hood.

In addition, add boundary integral to advection term for 'elev' bc this
makes P2P1 stable. This also affects other element pairs - need to
evaluate more.