Finish swe2d test and add new mms test.
The finished test is a steady state balance between pressure gradient
and a linear friction term. The reference solution neglects the advection
(non-linearity in divergence term is considered though).

The new "1D" MMS case (Method of Manufactured Solutions)
 again tests the solution in a steady state channel
case. This time a source term is added that exactly matches the
residual of a made up free surface solution - velocity is chosen such
that a constant flux is maintained consistent with the continuity
equation in a 1d channel. Errors are calculated from the
difference with the manufactured solution on a series of refined meshes.

For the RT(N+1)-PN element family we get roughly the expected N+1-order
convergence in both eta and u. The submitted test passes for N=1, also
tried N=0 and N=2 - but needs some tweaking of the tolerances and
timestepping settings for these to pass.

To enable direct solvers on the coupled system, switch to nest=False if
pc_type is set to lu.