test: More cache clearing in teardown
PyOP2 Kernel objects are cached on the class, but in the test suite we
make many of them, and they're /very/ big.  The complicated forms can
lead to Kernels that take up multiple megabytes of memory.  This can
quickly grow out of control.

"Fix" the problem by blowing away these class caches in the test
teardown.  The disk cache still exists, so we don't run through FFC
appreciably more, so hopefully everything is still fine.