Fix test and bc for ip viscosity.
Viscosity value was chosen too low to actually affect the convergence (for the meshes considered),
hiding the fact that the analytical source term was wrong. With a substantial viscosity however
we have to switch to implicit timestepping to avoid instability. There was also an error in the bc
for the full div-grad+grad-div form of viscosity.