Split the initialisation of FlowSolver2d in seperate stages:
* create_function_spaces()
* create_equations()
* create_timestepper()
* create_exporters()

Previously the last 3 where combined in a single create_equations()
call. However this meant you could not add terms to the equations before
the timestepper was set up. Being able to add terms to the equation from
outside the solver2d/shallow_water_equations modules, avoids unnecessary
code dependencies. For instance for tidal farms we need a drag term that
depends on (potentially) complicated thrust curves; sending all this
information down the FlowSolver2D -> ShallowWaterEquations ->
ShallowWaterEquationTerm path is cumbersome.

As before, if the user is not interested in fiddling with the equations,
there are checks to ensure the setup routines are called on the first
call to assign_initial_conditions() or iterate().