Use new integrated firedrake adjoint
This simplifies the code a lot as we no longer need a thetis_adjoint
thetis_config and a thetis/firedrake and no longer need to distinguish
between FiredrakeFunction (which was always the "real"
firedrake.Function and Function which was potentially the wrapped
firedrake_adjoint.Function).

Also removes a number of other workarounds (e.g. stop_annotation() around
split())