solver3d: fix bottom friction formulation
Bottom friction must be evaluated at the bottom surface to ensure that
velocity does not change sign near the bottom. As a consequence, the drag
coefficient cannot be computed directly with the law-of-the-wall equation.
Instead, it is estimated using an analytical linear fit to the log profile in
the bottom element.