Allow wetting and drying alpha to vary spatially (#186)
As it stands, the wetting and drying parameter which ensures a positive water height is assumed to be constant across the domain. Whilst this is a reasonable assumption for many problems, in some cases it would be advantageous to allow the parameter to vary spatially. One example is the case where we seek to model inundation of a coast near to a deep ocean trench with sharp gradients.