Don't use variant='equispaced' with RT
This was previously silently ignored, but now variant should be 'point'
(the old default), or 'integral' (soon to be new default). It looks like
we are failing three tests (2 in test_ekman_bottom, 1 in
test_atmospheric_pressure) with
variant='integral' so leaving it at variant='point'. This might just be
some tolerances that need tweaking.