Fix Thetis checkpoint loading
Previously we loaded the mesh from checkpoint, then loaded the field
into a temporary function based on that mesh, and then assigned it to
the field on the solver object, which is based on whatever mesh the user
provided us. This is no longer allowed (and also unsafe).

Now we ask the user to provide us with a mesh read from a checkpoint,
and directly read in the field into a function based on that same mesh.
To help the user achieve that added a read_mesh_from_checkpoint()
utility function (TBD).

See also discussion https://github.com/firedrakeproject/firedrake/discussions/3053

Should also fix #336