Fix NonHydrostaticTimeIntegrator2D for projection of solutions with adaptivity
The advance method of a timeintegrator should always advance from the
solved-for-fields, which are the same fields that will contain the
solution at the end of the timestep _after_ calling advance(). This is
because this is the only field(s) that the "user" of a time integrator
has control over, and the user may want to adjust/overwrite/project (for
adaptivity) at the end of a timestep and expect the timestepper to
advance from the adjusted values in the next timestep.

Thus, if a time-integrator stores some copy of a "solution_old" it
should be set at the _beginning_ of advance() to the current solution,
rather than at the end.