Add testing for demos, fixes #389
Tests are run in a similar way as for the examples. We only run them
to ensure they don't crash - in particular we want to spot errors due to
API changes - no checks for correctness.

This required some minor changes in north sea demo:
* move all input files to demos/runfiles
* add if os.environ['THETIS_REGRESSION_TEST'] logic to limit n/o
  timesteps, which should be hidden on web page
* add two dummy TPXO9 files so we can actually test the TPXO reader

The latter flagged an issue with uptide and the latest version of numpy
in which .tostring() (which is now .tobytes()) has been removed. Fixed
uptide version now on pypi.