Reusing Interpolator and don't use "work functions" in exporter
Previously we were using a "work function" (obtained via
function_space.get_work_function()) as the target for both the
projectors and interpolators used in the VTKEXporter. The way
we were using these was incredibly dodgy however, it relied
on use setting the max n/o work functions to 1 (which sets the
n/o of work functions to 1 to all functionspaces with the same
mesh and element, not just our local "fs_visu"!), and then lying
about releasing the work function, despite the work function
still be the target of the projector/interpolator, and then hoping
that the next call to get_work_function() we get the same one back!

For the Interpolator it doesn't really make sense any more, as it
doesn't actually hold a reference to a target function; you can
specify a target with tensor= at assembly, but we might as well just use
a temporary. For the projector, if we specify a target function space
instead of a target function, it will create a target function for us.
This does mean for the projector, we now store a separate target
function for each projector and thus for each specified output field. If
this is reallly a memory issue, then surely the bigger issue is that we
are creating separate interpolators and projectors _per_ output field
even if they are all on the same function space. Also, I don't think
we're using projection on output all that much in the first place
- it's off by default and not used in any tests or demos.